requirements.txt
apache-airflow==2.9.3
apache-airflow-providers-amazon==8.27.0
apache-airflow-providers-postgres==5.11.3
mlflow==3.10.0
flask<2.3
click<9
sqlalchemy<2
boto3==1.34.90
pandas==2.2.2
psycopg2-binary==2.9.9
scikit-learn==1.5.1
xgboost==2.1.3
imbalanced-learn==0.12.4
matplotlib==3.9.2